+++ src/AdminModel.java
@@ -3,7 +3,7 @@
 /**handles all calculations, database queries, and the overall work needed to be done for handling this associated role
  * Does NOT interact with the view class directly, and also does NOT interact with the Controller class(The Controller
  * class interacts with this class, not the other way around.)*/
-public class AdminModel extends UserModel {
+public class AdminModel extends UserSuperClass {