+++ src/Main.java
@@ -9,8 +9,8 @@ public class Main {
 	 * 		'Main.dbase' is used to access the object that represents the HashMap database.
 	 * 			ex. 'Main.dbase.containsKey()' checks the HashMap to see if a key is contained within it. */
 	
-	public static DatabaseModel dbaseClass = new DatabaseModel("dbase.txt");
-	public static HashMap<String, UserModel> dbase = dbaseClass.users;
+	public static Database dbaseClass = new Database("dbase.txt");
+	public static HashMap<String, UserSuperClass> dbase = dbaseClass.users;